LPS-79686 Splitting admin display contexts
