+++ modules/apps/collaboration/document-library/document-library-web/src/main/resources/META-INF/resources/document_library/navigation.jsp
@@ -18,11 +18,12 @@
 
 <%
 DLPortletInstanceSettingsHelper dlPortletInstanceSettingsHelper = new DLPortletInstanceSettingsHelper(dlRequestHelper);
+DLAdminNavigationDisplayContext dlAdminNavigationDisplayContext = new DLAdminNavigationDisplayContext(liferayPortletRequest, liferayPortletResponse);
 %>
 
 <c:if test="<%= dlPortletInstanceSettingsHelper.isShowTabs() %>">
 	<clay:navigation-bar
 		inverted="<%= true %>"
-		items="<%= dlAdminDisplayContext.getNavigationItems() %>"
+		items="<%= dlAdminNavigationDisplayContext.getNavigationItems() %>"
 	/>
 </c:if>
