+++ modules/apps/collaboration/document-library/document-library-web/src/main/resources/META-INF/resources/document_library/toolbar.jsp
@@ -20,25 +20,31 @@
 long repositoryId = GetterUtil.getLong((String)request.getAttribute("view.jsp-repositoryId"));
 
 long fileEntryTypeId = ParamUtil.getLong(request, "fileEntryTypeId", -1);
+
+DLAdminManagementToolbarDisplayContext dlAdminManagementToolbarDisplayContext = new DLAdminManagementToolbarDisplayContext(liferayPortletRequest, liferayPortletResponse, dlAdminDisplayContext);
 %>
 
 <clay:management-toolbar
-	actionItems="<%= dlAdminDisplayContext.getActionDropdownItems() %>"
-	clearResultsURL="<%= dlAdminDisplayContext.getClearResultsURL() %>"
-	creationMenu="<%= dlAdminDisplayContext.getCreationMenu() %>"
-	disabled="<%= dlAdminDisplayContext.isDisabled() %>"
-	filterItems="<%= dlAdminDisplayContext.getFilterDropdownItems() %>"
+	actionItems="<%= dlAdminManagementToolbarDisplayContext.getActionDropdownItems() %>"
+	clearResultsURL="<%= dlAdminManagementToolbarDisplayContext.getClearResultsURL() %>"
+	creationMenu="<%= dlAdminManagementToolbarDisplayContext.getCreationMenu() %>"
+	disabled="<%= dlAdminManagementToolbarDisplayContext.isDisabled() %>"
+	filterItems="<%= dlAdminManagementToolbarDisplayContext.getFilterDropdownItems() %>"
 	infoPanelId="infoPanelId"
-	searchActionURL="<%= String.valueOf(dlAdminDisplayContext.getSearchURL()) %>"
+	searchActionURL="<%= String.valueOf(dlAdminManagementToolbarDisplayContext.getSearchURL()) %>"
 	searchContainerId="entries"
-	selectable="<%= dlAdminDisplayContext.isSelectable() %>"
+	selectable="<%= dlAdminManagementToolbarDisplayContext.isSelectable() %>"
 	showInfoButton="<%= true %>"
-	showSearch="<%= dlAdminDisplayContext.isShowSearch() %>"
-	sortingOrder="<%= dlAdminDisplayContext.getOrderByType() %>"
-	sortingURL="<%= String.valueOf(dlAdminDisplayContext.getSortingURL()) %>"
-	totalItems="<%= dlAdminDisplayContext.getTotalItems() %>"
-	viewTypes="<%= dlAdminDisplayContext.getViewTypes() %>"
+	showSearch="<%= dlAdminManagementToolbarDisplayContext.isShowSearch() %>"
+	sortingOrder="<%= dlAdminManagementToolbarDisplayContext.getSortingOrder() %>"
+	sortingURL="<%= String.valueOf(dlAdminManagementToolbarDisplayContext.getSortingURL()) %>"
+	totalItems="<%= dlAdminManagementToolbarDisplayContext.getTotalItems() %>"
+	viewTypes="<%= dlAdminManagementToolbarDisplayContext.getViewTypes() %>"
 />
+
+<c:if test="<%= dlAdminManagementToolbarDisplayContext.isShowSearchInfo() %>">
+	<liferay-util:include page="/document_library/search_info.jsp" servletContext="<%= application %>" />
+</c:if>
 
 <aui:script>
 	function <portlet:namespace />deleteEntries() {
