+++ modules/apps/collaboration/document-library/document-library-web/src/main/resources/META-INF/resources/document_library/view.jsp
@@ -57,10 +57,6 @@ request.setAttribute("view.jsp-displayStyle", displayStyle);
 
 <liferay-util:include page="/document_library/toolbar.jsp" servletContext="<%= application %>" />
 
-<c:if test="<%= dlAdminDisplayContext.isShowSearchInfo() %>">
-	<liferay-util:include page="/document_library/search_info.jsp" servletContext="<%= application %>" />
-</c:if>
-
 <div id="<portlet:namespace />documentLibraryContainer">
 
 	<%
