+++ modules/apps/collaboration/document-library/document-library-web/src/main/resources/META-INF/resources/init.jsp
@@ -96,6 +96,8 @@ page import="com.liferay.document.library.kernel.util.VideoProcessorUtil" %><%@
 page import="com.liferay.document.library.web.internal.constants.DLWebKeys" %><%@
 page import="com.liferay.document.library.web.internal.dao.search.DLResultRowSplitter" %><%@
 page import="com.liferay.document.library.web.internal.dao.search.IGResultRowSplitter" %><%@
+page import="com.liferay.document.library.web.internal.display.context.DLAdminManagementToolbarDisplayContext" %><%@
+page import="com.liferay.document.library.web.internal.display.context.DLAdminNavigationDisplayContext" %><%@
 page import="com.liferay.document.library.web.internal.display.context.DLDisplayContextProvider" %><%@
 page import="com.liferay.document.library.web.internal.display.context.DLSelectRestrictedFileEntryTypesDisplayContext" %><%@
 page import="com.liferay.document.library.web.internal.display.context.DLViewFileEntryTypesDisplayContext" %><%@
